Move James Bond image to asset directory
